Statistik jumlah surat

+++ Surat/Surat/FormSuratDisposisiTambah.cs
@@ -91,7 +91,8 @@ namespace Surat
             else
             {
                 nama_gambar = "no_image.png";
-                File.Copy(Application.StartupPath + "\\no_image.png", lokasi_tujuan + "\\no_image.png", true);
+                if(!File.Exists(lokasi_tujuan + "\\no_image.png"))
+                    File.Copy(Application.StartupPath + "\\no_image.png", lokasi_tujuan + "\\no_image.png", true);
             }
 
             Database db = new Database();
